requirements.txt
langchain==0.2.12
langchain-community==0.2.10
langchain-openai==0.1.22
faiss-cpu==1.8.0
sentence-transformers==3.0.1
gradio==4.41.0
python-dotenv==1.0.1
openai==1.41.0
together==1.2.0
ollama==0.3.1
uvicorn==0.30.3
fastapi==0.111.1
# Databricks dependencies
databricks-sdk>=0.18.0
databricks-vectorsearch>=0.1.0
databricks-ai>=0.1.0
pyspark>=3.5.0
delta-spark>=3.0.0

